requirements.txt
beautifulsoup4
recipe-scrapers
pytesseract
Pillow
requests
markdown
markdownify
weasyprint
Jinja2
python-dotenv
